% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dartx.R
\name{phi}
\alias{phi}
\title{Estimate Values of Phi}
\usage{
phi(df, defaultPhi = TRUE, ...)
}
\arguments{
\item{df}{dataframe with a column cume_dist}

\item{...}{optional argument customPhi. Where customPhi is a dataframe with min, max, exp. Specifying the streamflow percentiles to apply values of phi (exp).}
}
\value{
dataframe
}
\description{
Estimate Values of Phi
}
\keyword{internal}
